Поддержка смены url для oauth
