+++ src/apollo/state/settings.query.graphql
@@ -1,0 +1,6 @@
+query getSettings {
+  appSettings @client {
+    url
+    groupName
+  }
+}
